requirements.txt
opencv-python~=4.5.2.52
Pillow~=8.2.0
Flask~=2.0.1
tensorflow~=2.5.0
numpy~=1.19.5
matplotlib~=3.4.2